zmiana nazw, user poza paczka menu, role dodane do konstruktora
