+++ src/main/java/Kantor/Currency.java
@@ -1,3 +1,5 @@
+package Kantor;
+
 import java.util.Date;
 
 public class Currency {
