+++ src/main/java/Kantor/Main.java
@@ -1,3 +1,5 @@
+package Kantor;
+
 public class Main {
     public static void main(String[] args) {
         Menu.getInstance().showMenu(Menu.MenuOptions.LOGIN_MENU);
