+++ src/main/java/Kantor/User.java
@@ -1,0 +1,29 @@
+package Kantor;
+
+public class User {
+    private String login;
+    private String password;
+    private Role role;
+
+    public User(String login, String password, Role role) {
+        this.login = login;
+        this.password = password;
+        this.role = role;
+    }
+
+    public String getLogin() {
+        return login;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public Role getRole() {
+        return role;
+    }
+
+    public enum Role{
+        USER, ADMIN;
+    }
+}
