+++ src/main/java/Kantor/menu/MenuInterface.java
@@ -1,4 +1,4 @@
-package optionsMenu;
+package Kantor.menu;
 
 public interface MenuInterface {
     public void display();
